[fixup] Refactor the process of generating NetworkEnvironmentConfiguraiton
